MDEV-31793 LooseScan join on derived table crashes
Statements that may be affected by this bug have
1) a semi join to a derived table
2) a text or blob field in the select list of the derived table
3) loose scan optimization enabled

When a temporary table is materialized that contains a BLOB or TEXT
field and a unique key key is added to this table (for any reason)
a new key field is added as the first key part. This field is to
differentiate NULL from blank strings.  This key part had it's store_length
set to zero, causing a potential loose scan to try and allocate then
free zero bytes.

Under normal circumstances, a loose scan is never performed on a derived
table, but it is possible to force it by either getting incorrect
statistics from the engine (see OP in MDEV) or altering the plan costs
and disabling some join types (see test case).
This loose scan has the potential to produce incorrect results, so we ban it.