MDEV-31949 slow parallel replication of user xa
This is the 2nd commit of the planned 3 commit series.
The first commit refactors the binary logging of XA to integrate
it properly with the Binlog-Group-Commit module.
The 3rd commit follows with an agenda to optimize away fdatasync
by the user XA-PREPARE.

Introduced by MDEV-742 XA-Prepare group of events

  XA START xid
  ...
  XA END xid
  XA PREPARE xid

and its XA-"complete" terminator

  XA COMMIT or
  XA ROLLBACK

that constitute a binlog image of a user XA transaction
have now been made distributed Round-Robin across slave parallel workers.
The original hash-based policy was perfect to identify the XA-PREPARE
worker by a respective XA-"COMPLETE".
Yet it was proven to attribute to execution latency through creating a
big - many times larger than the size of the worker pool - queue of
binlog-ordered transactions to commit.

Acronyms and notations used below:

  XAP := XA-Prepare event or the whole prepared XA group of events
  XAC := XA-"complete", which is a solitary group of events

Notable changes are done mostly to the parallel slave module.

The parallel slave driver thread now maintains a list XAP:s currently
in processing. It's purpose is to avoid "wild" parallel execution of XA:s
with duplicate xids (unlikely, but that's the user's right).
The list is arranged as a sliding window with the size of the slave
worker pool to account a possibility of the largest (in the
transaction count sense) dependency.
When a new XAP or XAC enters the replication applier its xid
is compared against ones in the window. If the current one duplicates,
e.g when it's XAC(xid) that comes in after XAP(xid), then
XAC would be 'waiting for prior transaction to commit'.
In this example among the prior transactions one would necessarily be
the XAC's "parent" XA-PREPARE part of the whole XA transaction.

Small changes done to
  ha_close_connection()
are done to enable it for the end of XA-prepare's binlogging
invocation on the parallel slave.

Thanks to Brandon Nesterenko at mariadb.com for initial review and
a lot of creative efforts to advance with this work!
Enormous kudos to Kristian Nielsen for thorough insightful review!
Roel Van de Paar helped to improved from the testing.