MDEV-31361 Wrong result on second execution of prepare statement for engine Federated with derived table
This issue was caused by a conflict between merging of a derived
table and pushing it down to the storage engine as a whole.
The "merged" flag is set at mysql_derived_init(), while the
pushdown handler is selected at a later phase,
at mysql_derived_prepare().

It does not make sense to set a pushdown handler for a
derived table which is going to be merged, so this fix disables
pushing down of such tables.