add workflow rules
Do not run pipelines on merge requests because this results in duplicate
pipelines. A pipeline will be triggered when a branch is pushed to and a
duplicate pipeline will be triggered when a merge request is created.

Do not run pipelines on the "gitlab-ci" branch since this branch is meant only
to perform RDS internal gitlab CI overrides.